ranking_system/requirements.txt
appdirs==1.4.3
atomicwrites==1.3.0
attrs==19.3.0
black==19.3b0
certifi==2019.9.11
chardet==3.0.4
Click==7.0
entrypoints==0.3
flake8==3.7.8
Flask==1.1.1
h5py==2.10.0
idna==2.8
importlib-metadata==0.23
itsdangerous==1.1.0
Jinja2==2.10.3
joblib==0.14.0
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
MarkupSafe==1.1.1
mccabe==0.6.1
more-itertools==7.2.0
numpy==1.17.3
packaging==19.2
pkg-resources==0.0.0
pluggy==0.13.0
py==1.8.0
pycodestyle==2.5.0
pyflakes==2.1.1
pymongo==3.9.0
pyparsing==2.4.2
pytest==5.2.2
PyYAML==5.1.2
requests==2.22.0
scikit-learn==0.21.3
scipy==1.3.1
six==1.12.0
sklearn==0.0
toml==0.10.0
urllib3==1.25.6
wcwidth==0.1.7
Werkzeug==0.16.0
zipp==0.6.0
